room: move world bounds calc to libtrx
