+++ src/libtrx/game/rooms/common.c
@@ -6,8 +6,7 @@
 #include "game/items.h"
 #include "game/rooms/const.h"
 #include "game/rooms/enum.h"
-
-#include <stddef.h>
+#include "utils.h"
 
 #define FD_NULL_INDEX 0
 #define FD_IS_DONE(t) ((t & 0x8000) == 0x8000)
@@ -213,4 +212,28 @@ int32_t Room_FindByPos(const int32_t x, const int32_t y, const int32_t z)
     }
 
     return NO_ROOM_NEG;
+}
+
+BOUNDS_32 Room_GetWorldBounds(void)
+{
+    BOUNDS_32 bounds = {
+        .min.x = 0x7FFFFFFF,
+        .min.z = 0x7FFFFFFF,
+        .max.x = 0,
+        .max.z = 0,
+        .min.y = MAX_HEIGHT,
+        .max.y = -MAX_HEIGHT,
+    };
+
+    for (int32_t i = 0; i < Room_GetTotalCount(); i++) {
+        const ROOM *const room = Room_Get(i);
+        bounds.min.x = MIN(bounds.min.x, room->pos.x);
+        bounds.max.x = MAX(bounds.max.x, room->pos.x + room->size.x * WALL_L);
+        bounds.min.z = MIN(bounds.min.z, room->pos.z);
+        bounds.max.z = MAX(bounds.max.z, room->pos.z + room->size.z * WALL_L);
+        bounds.min.y = MIN(bounds.min.y, room->max_ceiling);
+        bounds.max.y = MAX(bounds.max.y, room->min_floor);
+    }
+
+    return bounds;
 }
